requirements.txt
attrs==17.2.0
Automat==0.6.0
certifi==2017.4.17
chardet==3.0.4
constantly==15.1.0
falcon==1.2.0
future==0.16.0
geoip2==2.5.0
hiredis==0.2.0
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
maxminddb==1.3.0
netaddr==0.7.19
pyaml==16.12.2
python-mimeparse==1.6.0
PyYAML==3.12
redis==2.10.5
requests==2.18.1
six==1.10.0
Twisted==17.5.0
urllib3==1.21.1
zope.interface==4.4.2
